backend/requirements.txt
# ==================================
# DEPENDENCIAS BACKEND - Flask
# ==================================
# Versiones compatibles con Python 3.11 - 3.13

# Framework web
Flask==3.1.3
Werkzeug==3.1.6

# Base de datos y ORM
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48

# Migraciones de base de datos
Flask-Migrate==4.1.0
alembic==1.18.4

# Autenticación y seguridad
Flask-JWT-Extended==4.7.1
bcrypt==5.0.0

# CORS
Flask-CORS==6.0.2

# Validación y serialización
marshmallow==4.2.2

# Manejo de imágenes
Pillow==12.1.1

# Variables de entorno
python-dotenv==1.2.2

# Servidor WSGI para producción
gunicorn==25.1.0

# Utilidades
python-dateutil==2.9.0
pytz==2025.2

# Validación de formularios
email-validator==2.3.0

# Testing (opcional, para desarrollo)
pytest==9.0.2
pytest-flask==1.3.0
pytest-cov==7.0.0

# Linting y formato (opcional, para desarrollo)
flake8==7.3.0
black==26.3.0

# Generación de datos fake para testing (opcional)
Faker==40.8.0
